adjust search fields and results to append utms params
